backend/requirements.txt
anyio==4.2.0
aiohttp==3.9.1
bcrypt==4.1.2
fastapi==0.109.0
numpy==1.26.3
passlib==1.7.4
psycopg2-binary==2.9.9
pydantic==2.5.3
PyJWT==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.6
scipy==1.12.0
SQLAlchemy==2.0.25
flower==2.0.1
celery==5.3.6
redis==5.0.1
scikit-learn==1.4.0
requests==2.31.0
tenacity==8.2.3
uvicorn[standard]==0.27.0
gunicorn==21.2.0
eventlet==0.35.2
